Splits range unit test file into separate files for separate units
