+++ tiledb/sm/misc/test/unit_range_subset.cc
@@ -1,5 +1,5 @@
 /**
- * @file tiledb/sm/misc/test/unit_range.cc
+ * @file tiledb/sm/misc/test/unit_range_subset.cc
  *
  * @section LICENSE
  *
@@ -27,8 +27,8 @@
  *
  * @section DESCRIPTION
  *
- * This file defines unit tests for the Range class and helper functions that
- * operator on Ranges.
+ * This file defines unit tests for the Range helper functions used for subset
+ * comparisons.
  */
 
 #include <catch.hpp>
@@ -37,113 +37,6 @@
 using namespace tiledb;
 using namespace tiledb::common;
 using namespace tiledb::sm;
-
-/**************************/
-/* Test Check Valid Range */
-/**************************/
-template <typename T>
-void test_valid_range(T lower, T upper) {
-  T data[2]{lower, upper};
-  Range range{data, 2 * sizeof(T)};
-  auto status = check_typed_range_is_valid<T>(range);
-  REQUIRE(status.ok());
-}
-
-template <typename T>
-void test_invalid_range(T lower, T upper) {
-  T data[2]{lower, upper};
-  Range range{data, 2 * sizeof(T)};
-  auto status = check_typed_range_is_valid<T>(range);
-  REQUIRE(!status.ok());
-  INFO(status.ok())
-}
-
-template <typename T>
-void test_full_typeset_is_valid() {
-  T fullset[2]{std::numeric_limits<T>::min(), std::numeric_limits<T>::max()};
-  Range range{fullset, 2 * sizeof(T)};
-  auto status = check_typed_range_is_valid<T>(range);
-  REQUIRE(status.ok());
-}
-
-TEMPLATE_TEST_CASE(
-    "RangeOperations: Test is_valid_range for unsigned int types",
-    "[range]",
-    uint8_t,
-    uint16_t,
-    uint32_t,
-    uint64_t) {
-  SECTION("Test single point is valid") {
-    test_valid_range<TestType>(1, 1);
-  }
-  SECTION("Test standard range is valid") {
-    test_valid_range<TestType>(1, 10);
-  }
-  SECTION("Test full typeset is valid") {
-    test_full_typeset_is_valid<TestType>();
-  }
-  SECTION("Test lower bound larger than upper bound is invalid") {
-    test_invalid_range<TestType>(10, 1);
-  }
-}
-
-TEMPLATE_TEST_CASE(
-    "RangeOperations: Test is_valid_range for signed int types",
-    "[range]",
-    int8_t,
-    int16_t,
-    int32_t,
-    int64_t) {
-  SECTION("Test single point is valid") {
-    test_valid_range<TestType>(-1, -1);
-  }
-  SECTION("Test standard range is valid") {
-    test_valid_range<TestType>(-1, 10);
-  }
-  SECTION("Test full typeset is valid") {
-    test_full_typeset_is_valid<TestType>();
-  }
-  SECTION("Test lower bound larger than upper bound is invalid") {
-    test_invalid_range<TestType>(1, -1);
-  }
-}
-
-TEMPLATE_TEST_CASE(
-    "RangeOperations: Test is_valid_range for floating-point types",
-    "[range]",
-    float,
-    double) {
-  SECTION("Test single point is valid") {
-    test_valid_range<TestType>(1.0, 1.0);
-  }
-  SECTION("Test standard range is valid") {
-    test_valid_range<TestType>(-10.0, 10.0);
-  }
-  SECTION("Test full typeset is valid") {
-    test_full_typeset_is_valid<TestType>();
-  }
-  SECTION("Test lower bound larger than upper bound is invalid") {
-    test_invalid_range<TestType>(1.0, -1.0);
-  }
-  SECTION("Test infinite values are valid") {
-    test_valid_range<TestType>(
-        -std::numeric_limits<TestType>::infinity(),
-        std::numeric_limits<TestType>::infinity());
-    test_valid_range<TestType>(0.0, std::numeric_limits<TestType>::infinity());
-    test_valid_range<TestType>(-std::numeric_limits<TestType>::infinity(), 0.0);
-  }
-  SECTION("Test NaN values are invalid") {
-    test_invalid_range<double>(
-        std::numeric_limits<double>::quiet_NaN(),
-        std::numeric_limits<double>::quiet_NaN());
-    test_invalid_range<double>(0.0, std::numeric_limits<double>::quiet_NaN());
-    test_invalid_range<double>(std::numeric_limits<double>::quiet_NaN(), 0.0);
-  }
-}
-
-/****************************/
-/* Test Superset Operations */
-/****************************/
 
 template <typename T>
 void test_good_subset(const T* domain_data, const T* subset_data) {
